fix(shaple_flutter): add modules for mocking on test
