+++ packages/shaple_flutter/lib/shaple_flutter.dart
@@ -7,3 +7,4 @@ export 'src/flutter_go_true_client_options.dart';
 export 'src/local_storage.dart';
 export 'src/shaple.dart';
 export 'src/shaple_auth.dart';
+export 'src/stubs.dart';
