+++ packages/shaple_flutter/lib/src/stubs.dart
@@ -1,0 +1,73 @@
+import 'dart:convert';
+
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:shaple_flutter/shaple_flutter.dart';
+
+/// Construct session data for a given expiration date
+({String accessToken, String sessionString}) getSessionData(
+    DateTime accessTokenExpireDateTime) {
+  final accessTokenExpiresAt =
+      accessTokenExpireDateTime.millisecondsSinceEpoch ~/ 1000;
+  final accessTokenMid = base64.encode(utf8.encode(json.encode({
+    'exp': accessTokenExpiresAt,
+    'sub': '1234567890',
+    'role': 'authenticated'
+  })));
+  final accessToken = 'any.$accessTokenMid.any';
+  final sessionString =
+      '{"access_token":"$accessToken","expires_in":${accessTokenExpireDateTime.difference(DateTime.now()).inSeconds},"refresh_token":"-yeS4omysFs9tpUYBws9Rg","token_type":"bearer","provider_token":null,"provider_refresh_token":null,"user":{"id":"4d2583da-8de4-49d3-9cd1-37a9a74f55bd","app_metadata":{"provider":"email","providers":["email"]},"user_metadata":{"Hello":"World"},"aud":"","email":"fake1680338105@email.com","phone":"","created_at":"2023-04-01T08:35:05.208586Z","confirmed_at":null,"email_confirmed_at":"2023-04-01T08:35:05.220096086Z","phone_confirmed_at":null,"last_sign_in_at":"2023-04-01T08:35:05.222755878Z","role":"","updated_at":"2023-04-01T08:35:05.226938Z"}}';
+  return (accessToken: accessToken, sessionString: sessionString);
+}
+
+class MockLocalStorage extends LocalStorage {
+  @override
+  Future<void> initialize() async {}
+  @override
+  Future<String?> accessToken() async {
+    return getSessionData(DateTime.now().add(const Duration(hours: 1)))
+        .sessionString;
+  }
+
+  @override
+  Future<bool> hasAccessToken() async => true;
+  @override
+  Future<void> persistSession(String persistSessionString) async {}
+  @override
+  Future<void> removePersistedSession() async {}
+}
+
+/// Registers the mock handler for uni_links
+void mockAppLink({String? initialLink}) {
+  const channel = MethodChannel('com.llfbandit.app_links/messages');
+  const anotherChannel = MethodChannel('com.llfbandit.app_links/events');
+
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  // ignore: invalid_null_aware_operator
+  TestDefaultBinaryMessengerBinding.instance?.defaultBinaryMessenger
+      .setMockMethodCallHandler(channel, (call) async => initialLink);
+
+  // ignore: invalid_null_aware_operator
+  TestDefaultBinaryMessengerBinding.instance?.defaultBinaryMessenger
+      .setMockMethodCallHandler(anotherChannel, (message) async => null);
+}
+
+class MockAsyncStorage extends GotrueAsyncStorage {
+  final Map<String, String> _map = {};
+
+  @override
+  Future<String?> getItem({required String key}) async {
+    return _map[key];
+  }
+
+  @override
+  Future<void> removeItem({required String key}) async {
+    _map.remove(key);
+  }
+
+  @override
+  Future<void> setItem({required String key, required String value}) async {
+    _map[key] = value;
+  }
+}
